+++ CHANGELOG.md
@@ -2,8 +2,8 @@
 
 ## Unreleased
 
-- Added `outDrawColor()` and `outFillColor()` functions to
-  the `PdfDocument` class.
+- `PdfDocument`: Updated `outLineJoin()` function.
+- `PdfDocument`: Added `outDrawColor()` and `outFillColor()` functions.
 
 ## 3.0.8 - 2025-02-26